k-means: Minor changes in SQL/Python code to reflect changes in design document
Jira: None

Changes:
- In the k-means SQL/Python code, replaced "col_" by "expr_", because users are
  not limited to passing column names but any expression that has the correct
  type.
- Fixed a compiler warning in avg_vector.cpp.

+++ src/modules/linalg/avg_vector.cpp
@@ -109,7 +109,8 @@ avg_vector_transition::run(AnyType& args) {
     if (state.numRows == 0)
         state.initialize(*this, x.size());
     else if (x.size() != state.sumOfVectors.size()
-        || state.numDimensions != state.sumOfVectors.size())
+        || state.numDimensions !=
+            static_cast<uint32_t>(state.sumOfVectors.size()))
         throw std::invalid_argument("Invalid arguments: Dimensions of points "
             "not consistent.");
